[ALPS02964263] wlan:wfd add statistic feature for performance issue
[Detail]
  update __tmp define from u8 to u32 for u4LineScore in
  mtk_cfg80211_testmode_get_sta_statistics
[Solution]

MTK-Commit-Id: 300f3d7deec02f91aba2ee83ee9c115b1a21c192

CR-Id: ALPS02964263
Feature: p2p

+++ drivers/misc/mediatek/connectivity/wlan/gen4/os/linux/gl_cfg80211.c
@@ -1926,7 +1926,7 @@ mtk_cfg80211_testmode_get_sta_statistics(IN struct wiphy *wiphy, IN void *data, 
 	if (unlikely(nla_put(skb, NL80211_TESTMODE_STA_STATISTICS_MAC, MAC_ADDR_LEN, prParams->aucMacAddr) < 0))
 		goto nla_put_failure;
 	{
-		u8 __tmp = u4LinkScore;
+		u32 __tmp = u4LinkScore;
 
 		if (unlikely(nla_put(skb, NL80211_TESTMODE_STA_STATISTICS_LINK_SCORE, sizeof(u32), &__tmp) < 0))
 			goto nla_put_failure;
